pins, timing, paths (table above)

+++ core/src/storage/meta.rs
@@ -94,7 +94,7 @@ fn utc_to_epoch(year: i32, month: i32, day: i32, hour: i32, minute: i32) -> i64 
 fn epoch_to_label(epoch: i64) -> String {
     let z = epoch / 86400 + 719468;
     let era = (if z >= 0 { z } else { z - 146096 }) / 146097;
-    let doe = (z - era * 146097) as i64;
+    let doe = z - era * 146097;
     let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
     let y = (yoe + era * 400) as i32;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
